python/requirements.txt
openai>=1.0.0
pyaudio>=0.2.11
pyserial>=3.5
pygame>=2.5.0
